requirements.txt
Django>=4.2
djangorestframework
djangorestframework-simplejwt
psycopg2-binary
drf-spectacular
python-dotenv
pytz
openpyxl
pandas
django-cors-headers
gunicorn
